# backend/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de portafolio de inversiones personales\n\n## Autenticación\nEsta API utiliza JWT de Supabase Auth. Incluye el token en el header:\n`Authorization: Bearer \u003ctu-token-jwt\u003e`\n\n## Tipos de Activos Soportados\n- `stock` - Acciones\n- `etf` - Fondos cotizados\n- `bond` - Bonos\n- `crypto` - Criptomonedas\n- `real_estate` - Bienes raíces\n- `gold` - Oro\n- `cash` - Efectivo\n- `other` - Otros",
        "title": "WealthScope API",
        "contact": {
            "name": "WealthScope Team",
            "url": "https://github.com/Unikyri/WealthScope",
            "email": "support@wealthscope.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "wealthscope-production.up.railway.app",
    "basePath": "/",
    "paths": {
        "/api/v1/ai/briefing": {
            "get": {
                "description": "Generates AI-powered portfolio briefing with highlights, recommendations, and health score",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI Portfolio"
                ],
                "summary": "Get autonomous portfolio briefing",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.BriefingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.BriefingResponse": {
            "type": "object",
            "properties": {
                "alerts": {
                    "description": "Portfolio alerts",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.PortfolioAlert"
                    }
                },
                "briefing": {
                    "description": "AI-generated narrative",
                    "type": "string"
                },
                "generated_at": {
                    "description": "Timestamp",
                    "type": "string"
                },
                "health_score": {
                    "description": "Portfolio health",
                    "allOf": [
                        {
                            "$ref": "#/definitions/services.HealthScore"
                        }
                    ]
                },
                "highlights": {
                    "description": "Key events",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Highlight"
                    }
                },
                "recommendations": {
                    "description": "Action items",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "handlers.Highlight": {
            "type": "object",
            "properties": {
                "asset": {
                    "description": "Asset name or symbol",
                    "type": "string"
                },
                "change": {
                    "description": "\"+5.2%\" or \"Volatility increased\"",
                    "type": "string"
                },
                "message": {
                    "description": "Optional detailed message",
                    "type": "string"
                },
                "type": {
                    "description": "\"gain\", \"loss\", \"alert\"",
                    "type": "string"
                }
            }
        },
...
        "response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "request_id": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "services.AlertSeverity": {
            "type": "string",
            "enum": [
                "info",
                "warning",
                "critical"
            ],
            "x-enum-varnames": [
                "AlertSeverityInfo",
                "AlertSeverityWarning",
                "AlertSeverityCritical"
            ]
        },
        "services.AlertType": {
            "type": "string",
            "enum": [
                "price",
                "concentration",
                "correlation",
                "health",
                "performance"
            ],
            "x-enum-varnames": [
                "AlertTypePrice",
                "AlertTypeConcentration",
                "AlertTypeCorrelation",
                "AlertTypeHealth",
                "AlertTypePerformance"
            ]
        },
        "services.HealthScore": {
            "type": "object",
            "properties": {
                "correlation": {
                    "description": "10% weight - placeholder",
                    "type": "integer"
                },
                "diversification": {
                    "description": "25% weight",
                    "type": "integer"
                },
                "liquidity": {
                    "description": "10% weight",
                    "type": "integer"
                },
                "overall": {
                    "description": "0-100 weighted average",
                    "type": "integer"
                },
                "performance": {
                    "description": "15% weight",
                    "type": "integer"
                },
                "risk_level": {
                    "description": "20% weight",
                    "type": "integer"
                },
                "sector_balance": {
                    "description": "20% weight",
                    "type": "integer"
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "services.PortfolioAlert": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "asset": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "severity": {
                    "$ref": "#/definitions/services.AlertSeverity"
                },
                "type": {
                    "$ref": "#/definitions/services.AlertType"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ingresa tu JWT token con el prefijo \"Bearer \". Ejemplo: \"Bearer eyJhbGci...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}